sdk/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tx3-sdk"
version = "1.0.0"
description = "Tx3 SDK for Python"
readme = "README.md"
license = { text = "Apache-2.0" }
authors = [{ name = "TxPipe", email = "hello@txpipe.io" }]
requires-python = ">=3.10"
dependencies = [
  "bech32>=1.2.0",
  "bip-utils>=2.9.0",
  "httpx>=0.27.0",
  "pynacl>=1.5.0",
]

[project.urls]
Homepage = "https://github.com/tx3-lang/python-sdk"
Repository = "https://github.com/tx3-lang/python-sdk"
Issues = "https://github.com/tx3-lang/python-sdk/issues"

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.7",
  "ruff>=0.5.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
tx3_sdk = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
